requirements.txt
pandas
google-generativeai
supabase
python-dotenv

